requirements.txt
numpy
Pillow
matplotlib
tqdm
PyYAML
tensorboard
datasets
